+++ cypress/e2e/Ghost-5.27.0/staff/invite-staff-member.cy.js
@@ -1,21 +1,37 @@
-import { registerCommands } from '../../support/commands'
+import { registerCommands } from '../../../support/commands'
+import staffPage from '../../../pages/5.27.0/staffPage'
+import {faker} from '@faker-js/faker';
 
 registerCommands()
 
 let user = Cypress.config('user')
 let passw = Cypress.config('passw')
-
-describe ('Invite member staff', function(){
-    before(( ) => {
-        cy.loginAdmin(user,passw)
-        cy.url().should('include', '/dashboard')
-    })
-
-    it('Invite member', function(){
-        cy.visit(Cypress.env('url_staff'))
-        cy.contains('Invite people').click()
-        cy.url().should('include', '/invite')
-        cy.get(".peer").type('albertogalvis@protonmail.com')
-        cy.contains('Send invitation now').click()
-    });
+let emailUserInvited = faker.internet.email() 
+let urlVisit = Cypress.config('baseUrl') + Cypress.env('url_staff')
+before(() => {
+    //Login in Application
+    cy.loginAdmin(user,passw)
+    cy.url().should('include', '/dashboard')
+    cy.screenshot('5.27.0/staff/invite-staff-member.cy.js/login')
 });
+
+describe ('Invite staff member', function(){
+
+    it('P3: Invite staff member', function(){
+        // Given
+        cy.visit(urlVisit)
+        // When 
+        staffPage.btnInvite()
+        staffPage.typeInviteMemberOldVersion(emailUserInvited)
+        staffPage.sendInvitation()
+        // Then
+        cy.screenshot('5.27.0/staff/invite-staff-member.cy.js/P3-invite-staff-member') 
+    });
+
+});
+
+after(() => {
+    cy.logout()
+    cy.url().should('include', '/signin')
+    cy.screenshot('5.27.0/staff/invite-staff-member.cy.js/logout')
+});
