+++ cypress/e2e/Ghost-5.73.2/members/edit-member.cy.js
@@ -1,37 +1,52 @@
-import { registerCommands } from '../../../support/commands'
+import { registerCommands } from '../../../support/commands' 
+import memberPage from '../../../pages/memberPage'
+import {faker} from '@faker-js/faker';
 
 registerCommands()
 
 let user = Cypress.config('userv2')
-let passw = Cypress.config('passwv2')
+let passw = Cypress.config('passv2')
+let newuserEmail = faker.internet.email() 
 let version = "v2"
+let urlVisit = Cypress.config('baseUrlv2') + Cypress.env('url_members')
+before(() => {
+    //Login in Application
+    cy.loginAdmin(user,passw,version)
+    cy.url().should('include', '/dashboard')
+    cy.screenshot('5.73.2/members/edit-member.cy.js/login')
+});
+
 describe ('Edit members', function(){
 
-    before(( ) => {
-        cy.loginAdmin(user,passw,version)
-        cy.url().should('include', '/dashboard')
-    })    
-
-    const user = {
-        currentEmail: 'josebocanegra@uniandes.edu.co',
-        newEmail: 'jbnegra@gmail.com', 
-    };
-
-    it('edit current member', function(){
+    it('P2: edit current member', function(){
       
         cy.on('uncaught:exception', (err, runnable) => {
             return false
         })
-        cy.visit(Cypress.env('url_members'))
-        cy.contains(user.currentEmail).click()
+        // Given
+        cy.visit(urlVisit)
+        cy.wait(2000)
+        // When
+        //cy.contains(user.currentEmail).click()
+        memberPage.getMember()
         cy.wait(1000)
-        cy.getByTestInput('member-email').clear()
-        cy.getByTestInput('member-email').type(user.newEmail)
-        cy.contains('Save').click()
+        memberPage.clearEmailMember()
+        memberPage.typeNewUseremail(newuserEmail)
+        cy.wait(2000)
+        memberPage.saveMember()
+        cy.wait(2000)
+
+        // Then
+        cy.visit(urlVisit)
+        memberPage.visibleEmailMember(newuserEmail)
+        cy.screenshot('5.73.2/members/edit-member.cy.js/P2-edit-current-member')
+
     });
 
-    it('Verify edit member', function(){
-        cy.visit(Cypress.env('url_members'))
-        cy.contains(user.newEmail).should('be.visible')
-    });
 });
+
+after(() => {
+    cy.logout()
+    cy.url().should('include', '/signin')
+    cy.screenshot('5.73.2/members/edit-member.cy.js/logout')
+});
