+++ cypress/e2e/Ghost-5.73.2/staff/verify-comments-staff-member.cy.js
@@ -1,21 +1,40 @@
 import { registerCommands } from '../../../support/commands'
+import staffPage from '../../../pages/staffPage'
 
 registerCommands()
 
 let user = Cypress.config('userv2')
 let passw = Cypress.config('passwv2')
+let urlVisit = Cypress.config('baseUrlv2') + Cypress.env('url_staff')
+let urlVisit2 = Cypress.config('baseUrlv2') + Cypress.env('url_invite')
 let version = "v2"
-describe ('Verify enable comments notifications', function(){
-    before(( ) => {
-        cy.loginAdmin(user,passw,version)
-        cy.url().should('include', '/dashboard')
-    })
+before(() => {
+    //Login in Application
+    cy.loginAdmin(user,passw,version)
+    cy.url().should('include', '/dashboard')
+    cy.screenshot('5.73.2/staff/verify-comments-staff-member.cy.js/login')
+});
 
-    it('Enable from owner staff user', function(){
-        cy.visit(Cypress.env('url_staff'))
-        cy.getByTestId('owner-user').click()
+describe ('P4: Verify enable comments notifications', function(){
+
+    it('P4: Verify enable comments notifications', function(){
+        // Given
+        cy.wait(1000)
+        cy.visit(urlVisit)
         cy.wait(2000)
-        cy.get('input[type="checkbox"]').eq(0).should('be.checked')
-        cy.contains('Save & close').click()
+        // When 
+        staffPage.clickBtnProfile()
+        cy.wait(2000)
+        // Then
+        staffPage.verifyCheckedComments()
+        cy.screenshot('5.73.2/staff/verify-comments-staff-member.cy.js/P3-verify-enable-comments-notifications')
+        staffPage.saveProfile()
+        cy.contains('Done').click()
     });
 });
+
+after(() => {
+    cy.logout()
+    cy.url().should('include', '/signin')
+    cy.screenshot('5.73.2/staff/verify-comments-staff-member.cy.js/logout')
+});
