+++ cypress/fixtures/login.json
@@ -1,0 +1,11 @@
+[
+    {
+        "fullname":"Escenario exitoso"
+    },
+    {
+        "fullname":""
+    },
+    {
+        "fullname":"abcdefghijklmnñopqrstuvwxyzabcdefghijklmnñopqrstuvwxyzabcdefghijklmnñopqrstuvwxyzabcdefghijklmnñoabcdefghijklmnñopqrstuvwxyzabcdefghijklmnñopqrstuvwxyzabcdefghijklmnñopqrstuvwxyzabcdefghijklmnñoabcdefghijklmnñopqrstuvwxyzabcdefghijklmnñopqrstuvwxyzabcdefghijklmnñopqrstuvwxyzabcdefghijklmn"
+    }
+    ]
